List authors separately in book_page
Taking advantage of the BooksAuthorsLink table we can now
sort every single author as a unique entity; forgotten are the
days when we had to use a single, ugly blob containig all authors.

We can use the same process for the author_page, but so far
the sqlquery using a substring of the author name seems to
suffice.